MK-17: Change structure of example sentences
The strutcure of example sentences was changed from an array to a single object
that has the same structure as the previously existing example sentence
object. Removing the array removes the need for having multiple entries of
example sentences for one kanji. There is also no need to convert the string
input on the client side into a an array of example sentences.